Create Dockerfile #9
We use an official Python 3.9 image as the base image.
Environment variables are set to configure Python and the application's home directory.
We create a working directory named /app and set it as the current working directory.
The requirements.txt file is copied into the container, and we install the required Python packages.
The entire application code is copied into the container.
Port 8000 is exposed for the application to listen on.
The CMD instruction specifies the command to run when the container starts. In this example, it runs the Gunicorn web server to serve your web application (yourapp:app) on port 8000.
To use this Dockerfile:

Replace requirements.txt with the name of your project's requirements file.
Replace "yourapp:app" with the appropriate command to run your web application. Make sure to update it based on your project's structure.